backend/requirements.txt
fastapi
uvicorn
tensorflow
scikeras
pandas
numpy
joblib
scikit-learn
